Define macros to make using library easier.
Before, it could define defaults or eeprom but not both.
If both were needed, then initialization had to be done
twice, with same values.  This is now hidden with two
macros that do all the legwork.

Fixed bug in dump()/ddump() causing extra unrelated byte to be printed.

Fixed bug in eeprom_data() not accounting for starting address.

Updated all examples and readme.

Signed-off-by: Chris Evich <chris-arduino@anonomail.me>